Implement most of constructProlongation (compiles, but untested)

+++ include/gravomg/sampling.h
@@ -8,17 +8,17 @@
 namespace GravoMG {
 
 
-    std::vector<int> maximumDeltaIndependentSet(
+    std::vector<size_t> maximumDeltaIndependentSet(
             const Eigen::MatrixXd &pos, const Eigen::MatrixXi &edges, const double &radius
     );
 
-    std::vector<int> maximumDeltaIndependentSetWithDistances(
+    std::vector<size_t> maximumDeltaIndependentSetWithDistances(
             const Eigen::MatrixXd &pos, const Eigen::MatrixXi &edges,
             const double &radius,
             Eigen::VectorXd &D, std::vector<size_t> &nearestSourceK
     );
 
-    std::vector<int> fastDiscSample(const Eigen::MatrixXd &pos, const Eigen::MatrixXi &edges, const double &radius);
+    std::vector<size_t> fastDiscSample(const Eigen::MatrixXd &pos, const Eigen::MatrixXi &edges, const double &radius);
 }
 
 #endif //GRAVOMG_SAMPLING_H
